fix empty price field

+++ src/main/java/com/example/sql/proxy/dto/ItemDto.java
@@ -1,4 +1,6 @@
 package com.example.sql.proxy.dto;
 
-public record ItemDto(Integer id, String name, String sku, Integer price, String description) {
+import java.math.BigDecimal;
+
+public record ItemDto(Integer id, String name, String sku, BigDecimal price, String description) {
 }
